requirements.txt
Flask==0.10.1
gunicorn==19.4.5

alembic==0.8.6
Flask-Migrate==1.8.0

# database
Flask-SQLAlchemy===2.1
SQLAlchemy-Utils==0.32.6
psycopg2==2.6.1
passlib==1.6.5

PyYAML==3.11

# forms
Flask-WTF==0.12

# user session manager
Flask-Login==0.3.2

# apis
Flask-RESTful==0.3.5

# CORS
Flask-Cors==2.1.2

# Schema validation
marshmallow==2.7.3
marshmallow-jsonapi==0.7.1

# sending emails
Flask-SendGrid==0.1.0

# checkout
stripe==1.35.0
